# api-gateway/api/docs/swagger.yaml
definitions:
  genprotos.Candidate:
    properties:
      election:
        type: string
      id:
        type: string
      party:
        type: string
      public:
        type: string
    type: object
  genprotos.CandidateCreate:
    properties:
      ElectionId:
        type: string
      PartyId:
        type: string
      PublicId:
        type: string
    type: object
  genprotos.Election:
    properties:
      date:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  genprotos.GetAllCandidate:
    properties:
      Candidates:
        items:
          $ref: '#/definitions/genprotos.Candidate'
        type: array
    type: object
  genprotos.GetAllElection:
    properties:
      elections:
        items:
          $ref: '#/definitions/genprotos.Election'
        type: array
    type: object
...
  genprotos.Void:
    type: object
...
info:
  contact: {}
paths:
  /candidate:
    post:
      consumes:
      - application/json
      description: Creates candidate by reading from body
      operationId: create_candidate
      parameters:
      - description: candiate body data
        in: body
        name: Candiate
        required: true
        schema:
          $ref: '#/definitions/genprotos.CandidateCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Candiate created successfully
          schema:
            type: string
        "500":
          description: Failed to create candidate
          schema:
            type: string
      summary: Creates Candidate
      tags:
      - Candidate
  /candidate/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes candidate by reading id from body
      operationId: delete_candidate
      parameters:
      - description: candiate id data
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Candiate deleted successfully
          schema:
            type: string
        "500":
          description: Failed to create candidate
          schema:
            type: string
      summary: deletes Candidate
      tags:
      - Candidate
    get:
      consumes:
      - application/json
      description: Gets candidate by reading id from body
      operationId: get_candidate_by_id
      parameters:
      - description: candiate id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.Candidate'
        "500":
          description: Failed to get candidate by id
          schema:
            type: string
      summary: gets Candidate by id
      tags:
      - Candidate
  /candidate/by-party:
    get:
      consumes:
      - application/json
      description: Gets all candidates
      operationId: get_candidates_by_party
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllCandidate'
        "500":
          description: Failed to get candidates by party
          schema:
            type: string
      summary: gets Candidate by the party id
      tags:
      - Candidate
  /candidate/update:
    put:
      consumes:
      - application/json
      description: updates candidate by reading from body
      operationId: updates_candidate
      parameters:
      - description: candiate data
        in: body
        name: candiate
        required: true
        schema:
          $ref: '#/definitions/genprotos.Candidate'
      produces:
      - application/json
      responses:
        "200":
          description: Candiate updated successfully
          schema:
            type: string
        "500":
          description: Failed to create candidate
          schema:
            type: string
      summary: updates Candidate
      tags:
      - Candidate
  /candidates:
    get:
      consumes:
      - application/json
      description: Gets all candidates
      operationId: get_all_candidate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllCandidate'
        "500":
          description: Failed to get all candidates
          schema:
            type: string
      summary: gets all Candidate
      tags:
      - Candidate
  /election:
    post:
      consumes:
      - application/json
      description: Create election
      operationId: create_election
      parameters:
      - description: election body data
        in: body
        name: Election
        required: true
        schema:
          $ref: '#/definitions/genprotos.Election'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.Void'
        "500":
          description: Failed to get product by id
          schema:
            type: string
      summary: Creates Election
      tags:
      - Election
  /election/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes election by reading id
      operationId: delete_election
      parameters:
      - description: election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            type: string
        "500":
          description: Failed to delete election data
          schema:
            type: string
      summary: Deletes Election
      tags:
      - Election
    get:
      consumes:
      - application/json
      description: gets election by reading id  from path
      operationId: get_election_by_id
      parameters:
      - description: election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.Election'
        "500":
          description: Failed to get product by id
          schema:
            type: string
      summary: Gets Election by id
      tags:
      - Election
  /election/by-date:
    get:
      consumes:
      - application/json
      description: Gets Elections by date
      operationId: get_elections_by_date
      parameters:
      - description: election date
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllElection'
        "500":
          description: Failed to Get elections data
          schema:
            type: string
      summary: Getss Elections by date
      tags:
      - Election
  /election/update:
    put:
      consumes:
      - application/json
      description: Updates election by reading election from body
      operationId: update_election
      parameters:
      - description: election data
        in: body
        name: election
        required: true
        schema:
          $ref: '#/definitions/genprotos.Election'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.Void'
        "500":
          description: Failed to update election data
          schema:
            type: string
      summary: Update Election
      tags:
      - Election
  /elections:
    get:
      consumes:
      - application/json
      description: Gets all Elections
      operationId: get_all_elections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllElection'
        "500":
          description: Failed to Get all elections
          schema:
            type: string
      summary: Gets all Elections
      tags:
      - Election
...
swagger: "2.0"